feat: add initial CONTRIBUTING.md
Adds documentation for how to contribute to `xk6-coap`. This initial
documentation is intentionally lightweight and is expected to evolve
over time based on the scale of contributions. In addition to
documentation updates, it is also expected the the project will continue
to evolve to include more automated testing, allowing contributors and
maintainers to have a higher degree of confidence in the correctness of
proposed changes.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>